% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUtoarray.R
\name{addUtoarray}
\alias{addUtoarray}
\title{Adds unemployment rate to an array}
\usage{
addUtoarray(A, dime, uenames = c(u = "0", e = "1", r = "r"))
}
\arguments{
\item{A}{a_1 x ... x a_p}

\item{i}{an integer or a string indicating the dimention of the array that corresponds to the employed and unemployed}
}
\value{
uenames, a character string indicating the dimension names corresponding to employed,unemployed and rate( will be created)
}
\description{
Adds unemployment rate to an array
}
\examples{
A=array(1:(prod(2:5)),2:5);dimnames(A)[[2]]<-c("unemp","emp","nilf");names(dimnames(A))[2]<-"variable";addUtoarray(A,2,c("u"="unemp","e"="emp","r"="r"));addUtoarray(A,"variable",c("u"=1,"e"=2,"r"=4))
X=array(1:(prod(4:6)),4:6); "\%.\%"(W,X,j=2);
W\%.\%.\%.\%t(X);
}

