% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compMSE.R
\name{compMSE}
\alias{compMSE}
\title{Compute MSE}
\usage{
compMSE(A, Aref, averageindex = "i")
}
\arguments{
\item{A}{an array with names(dimnames(A))!=NULL}

\item{Aref}{an array with names(dimnames(Aref))!=NULL, and names(dimnames(Aref)) is a subset of names(dimnames(A));if Aref is null only variance and mean are reported}

\item{averageindex, }{a vector of strings, such that names(dimnames(Aref)) \cup averageindex\subset names(dimnames(A))}
}
\value{
MSE of A
}
\description{
Compute MSE
}
\examples{
A=array(1:(prod(2:6)),2:6);dimnames(A)<-list(s=1:2,m=1:3,y=c("1","0","_1","r"),e=1:5,i=1:6));
Aref=array(1:(prod(3:4)),3:4);dimnames(Aref)=dimnames(A)[2:3];
compMSE(A,Aref)
}

