% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.1.estimatesigma.R
\name{estimatesigma}
\alias{estimatesigma}
\title{estimatesigma(1,1,1,2,"false",syntheticcpspops,BB)}
\usage{
estimatesigma(s, i, m1, m2, b, syntheticcpspops, BB)
}
\description{
estimatesigma(1,1,1,2,"false",syntheticcpspops,BB)
}
