% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relMSE,R.R
\name{relMSE}
\alias{relMSE}
\title{Compute relative MSE}
\usage{
relMSE(MSEA, MSEAref)
}
\arguments{
\item{MSEA}{an array with names(dimnames(A))!=NULL}

\item{MSEAref}{an array with names(dimnames(Aref))!=NULL, and names(dimnames(Aref)) is a subset of names(dimnames(A));if Aref is null only variance and mean are reported}
}
\value{
MSEA/MSEAref
}
\description{
Compute relative MSE
}
\examples{
MSEA=array(1:(prod(2:6)),2:6);dimnames(MSEA)<-list(s=1:2,m=1:3,y=c("1","0","_1","r"),e=1:5,i=1:6);
MSEAref=array(1:(prod(3:4)),3:4);dimnames(MSEAref)=dimnames(MSEA)[2:3];
relMSE(MSEA,MSEAref)
}
