% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generation_functions.R
\name{sampleruleH}
\alias{sampleruleH}
\title{identical(samplerule2(1,1),samplerule(1,1:800,1))}
\usage{
sampleruleH(i, j, m)
}
\description{
identical(samplerule2(1,1),samplerule(1,1:800,1))
}
