% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generation_functions.R
\name{sampleruleS}
\alias{sampleruleS}
\title{Return atomic rotation group 1 for seed i (1000 possible values)}
\usage{
sampleruleS(i)
}
\arguments{
\item{i}{An integer between 1 and 1000}
}
\value{
The atomic rotation group 1 for seed i
}
\description{
Return atomic rotation group 1 for seed i (1000 possible values)
}
\examples{
sampleruleS(1,1:160,1)
identical(sampleruleS(i),samplerule(i,1:100,1))
identical((function(i,m){c(sapply((m+i-2)+c(1:4,13:16),sampleruleS))})(1,1),samplerule(1,1:800,1))
}
